GRIM/GROOVIE: Move StuffIt archive code to Common
