+++ engines/groovie/groovie.cpp
@@ -27,7 +27,6 @@
 #include "groovie/script.h"
 #include "groovie/music.h"
 #include "groovie/resource.h"
-#include "groovie/stuffit.h"
 #include "groovie/vdx.h"
 
 #ifdef ENABLE_GROOVIE2
@@ -39,6 +38,7 @@
 #include "common/events.h"
 #include "common/file.h"
 #include "common/macresman.h"
+#include "common/stuffit.h"
 #include "common/textconsole.h"
 
 #include "backends/audiocd/audiocd.h"
@@ -93,7 +93,7 @@ Common::Error GroovieEngine::run() {
 	if (_gameDescription->version == kGroovieV2 && getPlatform() == Common::kPlatformMacintosh) {
 		// Load the Mac installer with the lowest priority (in case the user has installed
 		// the game and has the MIDI folder present; faster to just load them)
-		Common::Archive *archive = createStuffItArchive("The 11th Hour Installer");
+		Common::Archive *archive = Common::createStuffItArchive("The 11th Hour Installer");
 
 		if (archive)
 			SearchMan.add("The 11th Hour Installer", archive);
